Move association transformers into own folder
