+++ src/Models/association/Transformers/ProductCategoryTransformer.ts
@@ -3,7 +3,7 @@ import ProductCategoryInterface from '../../../Interfaces/association/ProductCat
 import ProductInterface from '../../../Interfaces/product/Product';
 import ProductAssociationInterface from '../../../Interfaces/association/ProductAssociation';
 import ProductTransformer from '../../product/Transformers/ProductTransformer';
-import ProductAssociationTransformer from '../../product/Transformers/ProductAssociationTransformer';
+import ProductAssociationTransformer from './ProductAssociationTransformer';
 import BaseTransformer from '../../../BaseTransformer';
 
 export default class ProductCategoryTransformer extends BaseTransformer {
